Cleanups and add Analog Input
This cleans up the $cmd generation to make it consistant across all
methods.  It also extends the debugging $result generation on each of
the methods to closely match what you'd expect to get back.  The regex
for parsing the $result has been made more specific so as to be less
likely to return false information.  Finally, added get_analog, which
will return the values of an analog pin.